requirements.txt

aio-pika==9.5.2
aiohttp==3.9.1
aiohttp-retry==2.8.3

albumentations==1.3.1
alembic==1.14.1
altair==5.2.0
amqp==5.2.0


asyncpg==0.30.0
attrs==23.2.0
autopep8==2.0.4

barcode==1.0.4
bcrypt==4.2.1

docker==7.1.0

faiss-cpu==1.7.4
fastapi==0.115.0
fastjsonschema==2.19.0
fasttext==0.9.2
filelock==3.13.1
fire==0.7.0
Flask==2.2.5
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
gast==0.4.0
gdown==5.2.0
gensim==4.3.2
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
geventhttpclient==2.0.12
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.17.1
google-api-python-client==2.164.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-language==2.13.3
google-pasta==0.2.0
googleapis-common-protos==1.62.0
googlemaps==4.10.0
googletrans==4.0.0rc1
greenlet==3.0.2
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==23.0.0
h11==0.16.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2025.1.1
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==5.2.0
idna==2.10
lazy_loader==0.4
Levenshtein==0.25.1
libclang==16.0.6
llvmlite==0.41.1
locket==1.0.0
locust==2.25.0
lux-api==0.5.1
lux-widget==0.1.11
Mako==1.3.8
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.4
matplotlib-inline==0.1.6

ninja==1.11.1.3
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
notion-client==2.3.0
numba==0.58.1
numpy==1.23.1
oauthlib==3.2.2
ollama==0.4.8
openai==0.28.1

pamqp==3.3.0
pandas==1.5.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
passlib==1.7.4
patsy==0.5.4
phik==0.12.3
pickleshare==0.7.5
pika==1.3.2
Pillow==10.1.0
psycopg-binary==3.2.4
psycopg2==2.9.10


pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.17.2
PyGoogleVoice==0.5.2
PyJWT==2.8.0
pymongo==4.6.0
pyOpenSSL==23.3.0

python-barcode==0.15.1
python-bidi==0.6.6
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==1.0.0

python-json-logger==2.0.7

regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1

SQLAlchemy==2.0.37
sqlparse==0.4.4

uvicorn==0.31.0
validators==0.22.0
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==13.1

